+++ client/flutter/divkit/lib/src/core/widgets/div_tap_action_emitter.dart
@@ -1,6 +1,7 @@
 import 'package:divkit/src/core/action/action.dart';
 import 'package:divkit/src/core/action/action_converter.dart';
 import 'package:divkit/src/core/protocol/div_context.dart';
+import 'package:divkit/src/generated_sources/div_animation.dart';
 import 'package:divkit/src/generated_sources/generated_sources.dart' as dto;
 import 'package:divkit/src/utils/provider.dart';
 import 'package:divkit/src/utils/tap_builder.dart';
@@ -11,17 +12,20 @@ class DivTapActionModel with EquatableMixin {
   final List<DivAction> actions;
   final List<DivAction> longtapActions;
   final bool enabled;
+  final bool enabledAnimation;
 
   const DivTapActionModel({
     required this.actions,
     required this.longtapActions,
     required this.enabled,
+    required this.enabledAnimation,
   });
 
   static Stream<DivTapActionModel> from(
     BuildContext context,
     List<dto.DivAction> actions,
     List<dto.DivAction> longtapActions,
+    dto.DivAnimation? divAnimation,
   ) {
     final variables =
         DivKitProvider.watch<DivContext>(context)!.variableManager;
@@ -47,10 +51,14 @@ class DivTapActionModel with EquatableMixin {
         }
       }
 
+      final animationName =
+          await divAnimation?.name.resolveValue(context: context);
+
       return DivTapActionModel(
         actions: a,
         longtapActions: la,
         enabled: a.isNotEmpty || la.isNotEmpty,
+        enabledAnimation: !(animationName == DivAnimationName.noAnimation),
       );
     }).distinct();
   }
@@ -58,6 +66,8 @@ class DivTapActionModel with EquatableMixin {
   @override
   List<Object?> get props => [
         actions,
+        longtapActions,
+        enabledAnimation,
       ];
 }
 
@@ -65,6 +75,7 @@ class DivTapActionModel with EquatableMixin {
 class DivTapActionEmitter extends StatefulWidget {
   final List<dto.DivAction> actions;
   final List<dto.DivAction> longtapActions;
+  final DivAnimation? actionAnimation;
 
   final Widget child;
 
@@ -73,6 +84,7 @@ class DivTapActionEmitter extends StatefulWidget {
     required this.actions,
     required this.longtapActions,
     required this.child,
+    required this.actionAnimation,
   });
 
   @override
@@ -92,6 +104,7 @@ class _DivTapActionEmitterState extends State<DivTapActionEmitter> {
       context,
       widget.actions,
       widget.longtapActions,
+      widget.actionAnimation,
     );
   }
 
@@ -104,6 +117,7 @@ class _DivTapActionEmitterState extends State<DivTapActionEmitter> {
         context,
         widget.actions,
         widget.longtapActions,
+        widget.actionAnimation,
       );
     }
   }
@@ -130,6 +144,9 @@ class _DivTapActionEmitterState extends State<DivTapActionEmitter> {
                   }
                 },
                 builder: (context, pressed, hovered, child) {
+                  if (!model.enabledAnimation) {
+                    return child;
+                  }
                   double opacity = 1.0;
                   if (pressed) {
                     opacity = 0.6;