+++ client/flutter/divkit/lib/src/core/widgets/image/div_image_widget.dart
@@ -50,6 +50,7 @@ class _DivImageWidgetState extends State<DivImageWidget> {
         action: widget.data.action,
         actions: widget.data.actions,
         longtapActions: widget.data.longtapActions,
+        actionAnimation: widget.data.actionAnimation,
         child: StreamBuilder<DivImageModel>(
           stream: stream,
           builder: (context, snapshot) {