+++ client/flutter/divkit/lib/src/core/widgets/text/div_text_widget.dart
@@ -42,6 +42,7 @@ class _DivTextWidgetState extends State<DivTextWidget> {
         action: widget.data.action,
         actions: widget.data.actions,
         longtapActions: widget.data.longtapActions,
+        actionAnimation: widget.data.actionAnimation,
         child: StreamBuilder<DivTextModel>(
           stream: stream,
           builder: (context, snapshot) {